+++ crates/recursion/core/src/lib.rs
@@ -18,6 +18,7 @@ pub mod stark;
 pub mod sys;
 
 pub use runtime::*;
+pub use stark::new_vk_hash;
 
 // Re-export the stark stuff from `zkm_recursion_core` for now, until we will migrate it here.
 // pub use zkm_recursion_core::stark;
