+++ crates/recursion/core/src/stark/utils.rs
@@ -1,3 +1,10 @@
+use p3_bn254_fr::Bn254Fr;
+use p3_field::{FieldAlgebra, PrimeField32};
+use p3_symmetric::Permutation;
+use zkm_stark::PartStarkVerifyingKey;
+
+use crate::stark::{outer_perm, KoalaBearPoseidon2Outer};
+
 /// Returns whether the `ZKM_DEV` environment variable is enabled or disabled.
 ///
 /// This variable controls whether a smaller version of the circuit will be used for generating the
@@ -27,4 +34,17 @@ pub fn zkm_common_mode() -> bool {
         tracing::warn!("ZKM_COMMON environment variable is enabled.");
     }
     enabled
+}
+
+/// If zkm_common_mode() returns true, combine the base vkey hash with `vk_commitment` and `pc_start`
+/// using a Poseidon2 permutation.
+pub fn new_vk_hash(
+    vk: &PartStarkVerifyingKey<KoalaBearPoseidon2Outer>,
+    vkey_hash: Bn254Fr,
+) -> Bn254Fr {
+    let commitment: [Bn254Fr; 1] = vk.commit.into();
+    let pc_start_bn254 = Bn254Fr::from_canonical_u32(vk.pc_start.as_canonical_u32());
+    let mut state = [vkey_hash, commitment[0], pc_start_bn254];
+    outer_perm().permute_mut(&mut state);
+    state[0]
 }
